Everyauth addition
Adding everyauth:

NOTE: you need to add the following to your /etc/hosts until we get the domain up….

127.0.0.1	sharepodz.com

Also, for now until the user is added to the db clicking on the login link on the home page will crash the server. To resolve this store the info coming back from the user into the db.

Signed-off-by: Mark Hazlett <mark@icog-studios.com>